+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Below is a historical runthrough of changes.
 
+# 0.1.1
+
+- Added `linux.rb` recipe that installs needed packages and mounts NFS shares
+- Creates correct folder and mounts `/disks/Torrents`
+
 # 0.1.0
 
 Initial release.